% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ClusterMetrics.hdfsMetrics}
\alias{ClusterMetrics.hdfsMetrics}
\title{ClusterMetrics.hdfsMetrics Object}
\usage{
ClusterMetrics.hdfsMetrics()
}
\value{
ClusterMetrics.hdfsMetrics object
}
\description{
ClusterMetrics.hdfsMetrics Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The HDFS metrics.
}
\seealso{
Other ClusterMetrics functions: \code{\link{ClusterMetrics.yarnMetrics}},
  \code{\link{ClusterMetrics}}
}
