% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{DiagnoseClusterOutputLocation}
\alias{DiagnoseClusterOutputLocation}
\title{DiagnoseClusterOutputLocation Object}
\usage{
DiagnoseClusterOutputLocation(outputUri = NULL)
}
\arguments{
\item{outputUri}{Output-only The Google Cloud Storage URI of the diagnostic output}
}
\value{
DiagnoseClusterOutputLocation object
}
\description{
DiagnoseClusterOutputLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The location of diagnostic output.
}
