% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{DiagnoseClusterResults}
\alias{DiagnoseClusterResults}
\title{DiagnoseClusterResults Object}
\usage{
DiagnoseClusterResults(outputUri = NULL)
}
\arguments{
\item{outputUri}{[Output-only] The Google Cloud Storage URI of the diagnostic output}
}
\value{
DiagnoseClusterResults object
}
\description{
DiagnoseClusterResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The location of diagnostic output.
}

