% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{GceClusterConfiguration}
\alias{GceClusterConfiguration}
\title{GceClusterConfiguration Object}
\usage{
GceClusterConfiguration(GceClusterConfiguration.metadata = NULL,
  tags = NULL, serviceAccount = NULL, subnetworkUri = NULL,
  networkUri = NULL, zoneUri = NULL, internalIpOnly = NULL,
  metadata = NULL, serviceAccountScopes = NULL)
}
\arguments{
\item{GceClusterConfiguration.metadata}{The \link{GceClusterConfiguration.metadata} object or list of objects}

\item{tags}{The Google Compute Engine tags to add to all instances}

\item{serviceAccount}{Optional The service account of the instances}

\item{subnetworkUri}{The Google Compute Engine subnetwork to be used for machine communications}

\item{networkUri}{The Google Compute Engine network to be used for machine communications}

\item{zoneUri}{Required The zone where the Google Compute Engine cluster will be located}

\item{internalIpOnly}{If true, all instances in the cluser will only have internal IP addresses}

\item{metadata}{The Google Compute Engine metadata entries to add to all instances}

\item{serviceAccountScopes}{The URIs of service account scopes to be included in Google Compute Engine instances}
}
\value{
GceClusterConfiguration object
}
\description{
GceClusterConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common configuration settings for resources of Google Compute Engine cluster instances, applicable to all instances in the cluster.
}
\seealso{
Other GceClusterConfiguration functions: \code{\link{GceClusterConfiguration.metadata}}
}
