% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{HadoopJob}
\alias{HadoopJob}
\title{HadoopJob Object}
\usage{
HadoopJob(HadoopJob.properties = NULL, mainJarFileUri = NULL,
  mainClass = NULL, args = NULL, jarFileUris = NULL, fileUris = NULL,
  archiveUris = NULL, properties = NULL, loggingConfiguration = NULL)
}
\arguments{
\item{HadoopJob.properties}{The \link{HadoopJob.properties} object or list of objects}

\item{mainJarFileUri}{The Hadoop Compatible Filesystem (HCFS) URI of the jar file containing the main class}

\item{mainClass}{The name of the driver's main class}

\item{args}{[Optional] The arguments to pass to the driver}

\item{jarFileUris}{[Optional] Jar file URIs to add to the CLASSPATHs of the Hadoop driver and tasks}

\item{fileUris}{[Optional] HCFS URIs of files to be copied to the working directory of Hadoop drivers and distributed tasks}

\item{archiveUris}{[Optional] HCFS URIs of archives to be extracted in the working directory of Hadoop drivers and tasks}

\item{properties}{[Optional] A mapping of property names to values, used to configure Hadoop}

\item{loggingConfiguration}{[Optional] The runtime log configuration for job execution}
}
\value{
HadoopJob object
}
\description{
HadoopJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud Dataproc job for running Hadoop MapReduce jobs on YARN.
}
\seealso{
Other HadoopJob functions: \code{\link{HadoopJob.properties}}
}

