% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{HiveJob.properties}
\alias{HiveJob.properties}
\title{HiveJob.properties Object}
\usage{
HiveJob.properties()
}
\value{
HiveJob.properties object
}
\description{
HiveJob.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] A mapping of property names and values, used to configure Hive. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/hadoop/conf/*-site.xml, /etc/hive/conf/hive-site.xml, and classes in user code.
}
\seealso{
Other HiveJob functions: \code{\link{HiveJob.scriptVariables}},
  \code{\link{HiveJob}}
}

