% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{HiveJob}
\alias{HiveJob}
\title{HiveJob Object}
\usage{
HiveJob(HiveJob.scriptVariables = NULL, HiveJob.properties = NULL,
  queryFileUri = NULL, queryList = NULL, continueOnFailure = NULL,
  scriptVariables = NULL, properties = NULL, jarFileUris = NULL)
}
\arguments{
\item{HiveJob.scriptVariables}{The \link{HiveJob.scriptVariables} object or list of objects}

\item{HiveJob.properties}{The \link{HiveJob.properties} object or list of objects}

\item{queryFileUri}{The HCFS URI of the script that contains Hive queries}

\item{queryList}{A list of queries}

\item{continueOnFailure}{[Optional] Whether to continue executing queries if a query fails}

\item{scriptVariables}{[Optional] Mapping of query variable names to values (equivalent to the Hive command: `SET name='value';`)}

\item{properties}{[Optional] A mapping of property names and values, used to configure Hive}

\item{jarFileUris}{[Optional] HCFS URIs of jar files to add to the CLASSPATH of the Hive server and Hadoop MapReduce (MR) tasks}
}
\value{
HiveJob object
}
\description{
HiveJob Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud Dataproc job for running Hive queries on YARN.
}
\seealso{
Other HiveJob functions: \code{\link{HiveJob.properties}},
  \code{\link{HiveJob.scriptVariables}}
}

