% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{JobPlacement}
\alias{JobPlacement}
\title{JobPlacement Object}
\usage{
JobPlacement(clusterName = NULL, clusterUuid = NULL)
}
\arguments{
\item{clusterName}{Required The name of the cluster where the job will be submitted}

\item{clusterUuid}{Output-only A cluster UUID generated by the Dataproc service when the job is submitted}
}
\value{
JobPlacement object
}
\description{
JobPlacement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Dataproc job configuration.
}
