% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{JobScheduling}
\alias{JobScheduling}
\title{JobScheduling Object}
\usage{
JobScheduling(maxFailuresPerHour = NULL)
}
\arguments{
\item{maxFailuresPerHour}{Optional Maximum number of times per hour a driver may be restarted as a result of driver terminating with non-zero code before job is reported failed}
}
\value{
JobScheduling object
}
\description{
JobScheduling Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Job scheduling options.Beta Feature: These options are available for testing purposes only. They may be changed before final release.
}
