% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{LoggingConfiguration.driverLogLevels}
\alias{LoggingConfiguration.driverLogLevels}
\title{LoggingConfiguration.driverLogLevels Object}
\usage{
LoggingConfiguration.driverLogLevels()
}
\value{
LoggingConfiguration.driverLogLevels object
}
\description{
LoggingConfiguration.driverLogLevels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The per-package log levels for the driver. This may include 'root' package name to configure rootLogger. Examples: 'com.google = FATAL', 'root = INFO', 'org.apache = DEBUG'
}
\seealso{
Other LoggingConfiguration functions: \code{\link{LoggingConfiguration}}
}

