% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{LoggingConfiguration}
\alias{LoggingConfiguration}
\title{LoggingConfiguration Object}
\usage{
LoggingConfiguration(LoggingConfiguration.driverLogLevels = NULL,
  driverLogLevels = NULL)
}
\arguments{
\item{LoggingConfiguration.driverLogLevels}{The \link{LoggingConfiguration.driverLogLevels} object or list of objects}

\item{driverLogLevels}{The per-package log levels for the driver}
}
\value{
LoggingConfiguration object
}
\description{
LoggingConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The runtime logging configuration of the job.
}
\seealso{
Other LoggingConfiguration functions: \code{\link{LoggingConfiguration.driverLogLevels}}
}
