% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{ManagedGroupConfiguration}
\alias{ManagedGroupConfiguration}
\title{ManagedGroupConfiguration Object}
\usage{
ManagedGroupConfiguration(instanceTemplateName = NULL,
  instanceGroupManagerName = NULL)
}
\arguments{
\item{instanceTemplateName}{[Output-only] The name of the Instance Template used for the Managed Instance Group}

\item{instanceGroupManagerName}{[Output-only] The name of the Instance Group Manager for this group}
}
\value{
ManagedGroupConfiguration object
}
\description{
ManagedGroupConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the resources used to actively manage an instance group.
}

