% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{NodeInitializationAction}
\alias{NodeInitializationAction}
\title{NodeInitializationAction Object}
\usage{
NodeInitializationAction(executableFile = NULL, executionTimeout = NULL)
}
\arguments{
\item{executableFile}{Required Google Cloud Storage URI of executable file}

\item{executionTimeout}{Optional Amount of time executable has to complete}
}
\value{
NodeInitializationAction object
}
\description{
NodeInitializationAction Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies an executable to run on a fully configured node and a timeout period for executable completion.
}
