% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{OperationMetadata}
\alias{OperationMetadata}
\title{OperationMetadata Object}
\usage{
OperationMetadata(operationType = NULL, description = NULL,
  warnings = NULL, status = NULL, statusHistory = NULL,
  clusterName = NULL, clusterUuid = NULL)
}
\arguments{
\item{operationType}{Output-only The operation type}

\item{description}{Output-only Short description of operation}

\item{warnings}{Output-only Errors encountered during operation execution}

\item{status}{Output-only Current operation status}

\item{statusHistory}{Output-only Previous operation status}

\item{clusterName}{Name of the cluster for the operation}

\item{clusterUuid}{Cluster UUId for the operation}
}
\value{
OperationMetadata object
}
\description{
OperationMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata describing the operation.
}
