% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{PigJob.properties}
\alias{PigJob.properties}
\title{PigJob.properties Object}
\usage{
PigJob.properties()
}
\value{
PigJob.properties object
}
\description{
PigJob.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] A mapping of property names to values, used to configure Pig. Properties that conflict with values set by the Cloud Dataproc API may be overwritten. Can include properties set in /etc/hadoop/conf/*-site.xml, /etc/pig/conf/pig.properties, and classes in user code.
}
\seealso{
Other PigJob functions: \code{\link{PigJob.scriptVariables}},
  \code{\link{PigJob}}
}

