% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.clusters.create}
\alias{projects.clusters.create}
\title{Creates a cluster in a project.}
\usage{
projects.clusters.create(Cluster, projectId)
}
\arguments{
\item{Cluster}{The \link{Cluster} object to pass to this method}

\item{projectId}{Required The ID of the Google Cloud Platform project that the cluster belongs to}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}

Other Cluster functions: \code{\link{Cluster.labels}},
  \code{\link{Cluster}},
  \code{\link{projects.clusters.patch}}
}
