% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.clusters.diagnose}
\alias{projects.clusters.diagnose}
\title{Gets cluster diagnostic information. After the operation completes, the Operation.response field contains DiagnoseClusterOutputLocation.}
\usage{
projects.clusters.diagnose(DiagnoseClusterRequest, projectId, clusterName)
}
\arguments{
\item{DiagnoseClusterRequest}{The \link{DiagnoseClusterRequest} object to pass to this method}

\item{projectId}{Required The ID of the Google Cloud Platform project that the cluster belongs to}

\item{clusterName}{Required The cluster name}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}

Other DiagnoseClusterRequest functions: \code{\link{DiagnoseClusterRequest}}
}
