% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.clusters.get}
\alias{projects.clusters.get}
\title{Gets the resource representation for a cluster in a project.}
\usage{
projects.clusters.get(projectId, clusterName)
}
\arguments{
\item{projectId}{[Required] The ID of the Google Cloud Platform project that the cluster belongs to}

\item{clusterName}{[Required] The cluster name}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}
}

