% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.clusters.list}
\alias{projects.clusters.list}
\title{Lists all clusters in a project.}
\usage{
projects.clusters.list(projectId, filter = NULL, pageToken = NULL,
  pageSize = NULL)
}
\arguments{
\item{projectId}{Required The ID of the Google Cloud Platform project that the cluster belongs to}

\item{filter}{Optional A filter constraining which clusters to list}

\item{pageToken}{The standard List page token}

\item{pageSize}{The standard List page size}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}
}
