% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.jobs.list}
\alias{projects.jobs.list}
\title{Lists jobs in a project.}
\usage{
projects.jobs.list(projectId, pageSize = NULL, pageToken = NULL,
  clusterName = NULL, jobStateMatcher = NULL)
}
\arguments{
\item{projectId}{[Required] The ID of the Google Cloud Platform project that the job belongs to}

\item{pageSize}{[Optional] The number of results to return in each response}

\item{pageToken}{[Optional] The page token, returned by a previous call, to request the next page of results}

\item{clusterName}{[Optional] If set, the returned jobs list includes only jobs that were submitted to the named cluster}

\item{jobStateMatcher}{[Optional] Specifies enumerated categories of jobs to list}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}
}

