% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_functions.R
\name{projects.jobs.patch}
\alias{projects.jobs.patch}
\title{Updates a job in a project.}
\usage{
projects.jobs.patch(Job, projectId, jobId, updateMask = NULL)
}
\arguments{
\item{Job}{The \link{Job} object to pass to this method}

\item{projectId}{Required The ID of the Google Cloud Platform project that the job belongs to}

\item{jobId}{Required The job ID}

\item{updateMask}{Required Specifies the path, relative to <code>Job</code>, of the field to update}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/dataproc/}{Google Documentation}

Other Job functions: \code{\link{Job.labels}},
  \code{\link{Job}}
}
