% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{QueryList}
\alias{QueryList}
\title{QueryList Object}
\usage{
QueryList(queries = NULL)
}
\arguments{
\item{queries}{Required The queries to execute}
}
\value{
QueryList object
}
\description{
QueryList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of queries to run on a cluster.
}
