% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SoftwareConfiguration.properties}
\alias{SoftwareConfiguration.properties}
\title{SoftwareConfiguration.properties Object}
\usage{
SoftwareConfiguration.properties()
}
\value{
SoftwareConfiguration.properties object
}
\description{
SoftwareConfiguration.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] The properties to set on daemon configuration files. Property keys are specified in 'prefix:property' format, such as 'core:fs.defaultFS'. The following are supported prefixes and their mappings: core - core-site.xml hdfs - hdfs-site.xml mapred - mapred-site.xml yarn - yarn-site.xml hive - hive-site.xml pig - pig.properties spark - spark-defaults.conf
}
\seealso{
Other SoftwareConfiguration functions: \code{\link{SoftwareConfiguration}}
}

