% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SoftwareConfiguration}
\alias{SoftwareConfiguration}
\title{SoftwareConfiguration Object}
\usage{
SoftwareConfiguration(SoftwareConfiguration.properties = NULL,
  imageVersion = NULL, properties = NULL)
}
\arguments{
\item{SoftwareConfiguration.properties}{The \link{SoftwareConfiguration.properties} object or list of objects}

\item{imageVersion}{Optional The version of software inside the cluster}

\item{properties}{Optional The properties to set on daemon configuration files}
}
\value{
SoftwareConfiguration object
}
\description{
SoftwareConfiguration Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the selection and configuration of software inside the cluster.
}
\seealso{
Other SoftwareConfiguration functions: \code{\link{SoftwareConfiguration.properties}}
}
