% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{SparkSqlJob.properties}
\alias{SparkSqlJob.properties}
\title{SparkSqlJob.properties Object}
\usage{
SparkSqlJob.properties()
}
\value{
SparkSqlJob.properties object
}
\description{
SparkSqlJob.properties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
[Optional] A mapping of property names to values, used to configure Spark SQL's SparkConf. Properties that conflict with values set by the Cloud Dataproc API may be overwritten.
}
\seealso{
Other SparkSqlJob functions: \code{\link{SparkSqlJob.scriptVariables}},
  \code{\link{SparkSqlJob}}
}

