% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{Job}
\alias{Job}
\title{Job Object}
\usage{
Job(reference = NULL, placement = NULL, hadoopJob = NULL,
  sparkJob = NULL, pysparkJob = NULL, hiveJob = NULL, pigJob = NULL,
  sparkSqlJob = NULL, status = NULL, statusHistory = NULL,
  yarnApplications = NULL, submittedBy = NULL,
  driverInputResourceUri = NULL, driverOutputResourceUri = NULL,
  driverControlFilesUri = NULL, interactive = NULL)
}
\arguments{
\item{reference}{[Optional] The fully qualified reference to the job, which can be used to obtain the equivalent REST path of the job resource}

\item{placement}{[Required] Job information, including how, when, and where to run the job}

\item{hadoopJob}{Job is a Hadoop job}

\item{sparkJob}{Job is a Spark job}

\item{pysparkJob}{Job is a Pyspark job}

\item{hiveJob}{Job is a Hive job}

\item{pigJob}{Job is a Pig job}

\item{sparkSqlJob}{Job is a SparkSql job}

\item{status}{[Output-only] The job status}

\item{statusHistory}{[Output-only] The previous job status}

\item{yarnApplications}{[Output-only] The collection of YARN applications spun up by this job}

\item{submittedBy}{[Output-only] The email address of the user submitting the job}

\item{driverInputResourceUri}{[Output-only] A URI pointing to the location of the stdin of the job's driver program, only set if the job is interactive}

\item{driverOutputResourceUri}{[Output-only] A URI pointing to the location of the stdout of the job's driver program}

\item{driverControlFilesUri}{[Output-only] If present, the location of miscellaneous control files which may be used as part of job setup and handling}

\item{interactive}{[Optional] If set to `true`, the driver's stdin will be kept open and `driver_input_uri` will be set to provide a path at which additional input can be sent to the driver}
}
\value{
Job object
}
\description{
Job Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cloud Dataproc job resource.
}

