% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc_objects.R
\name{JobStatus}
\alias{JobStatus}
\title{JobStatus Object}
\usage{
JobStatus(state = NULL, details = NULL, stateStartTime = NULL)
}
\arguments{
\item{state}{[Required] A state message specifying the overall job state}

\item{details}{[Optional] Job state details, such as an error description if the state is ERROR}

\item{stateStartTime}{[Output-only] The time when this state was entered}
}
\value{
JobStatus object
}
\description{
JobStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Cloud Dataproc job status.
}

